+++ src/scripts.h
@@ -149,13 +149,13 @@ typedef struct Script {
 
 extern const char* gScriptProcNames[SCRIPT_PROC_COUNT];
 
-unsigned int gameTimeGetTime();
+uint64_t gameTimeGetTime();
 void gameTimeGetDate(int* monthPtr, int* dayPtr, int* yearPtr);
 int gameTimeGetHour();
 char* gameTimeGetTimeString();
 void gameTimeAddTicks(int a1);
 void gameTimeAddSeconds(int a1);
-void gameTimeSetTime(unsigned int time);
+void gameTimeSetTime(uint64_t time);
 int gameTimeScheduleUpdateEvent();
 int gameTimeEventProcess(Object* obj, void* data);
 int _scriptsCheckGameEvents(int* moviePtr, int window);
